Task(#33): Implement build main feed use case
Includes:
- Some package refactor
- Some class name refactor
- MainFeedItem.kt created
