+++ app/src/main/java/com/jsmirabal/animeinfo/domain/model/DomainAnimeDetail.kt
@@ -1,5 +1,8 @@
 package com.jsmirabal.animeinfo.domain.model
 
+import com.jsmirabal.animeinfo.domain.model.base.AnimeDefinition
+import com.jsmirabal.animeinfo.domain.model.base.AnimeExtension
+
 data class DomainAnimeDetail(
     private val definition: AnimeDefinition,
     private val extension: AnimeExtension,
