+++ app/src/main/java/com/jsmirabal/animeinfo/domain/model/DomainTopAnimes.kt
@@ -1,3 +1,10 @@
 package com.jsmirabal.animeinfo.domain.model
 
-data class DomainTopAnimes(val topAnimes: List<AnimeCompact>)
+import com.jsmirabal.animeinfo.domain.model.base.AnimeCompact
+import com.jsmirabal.animeinfo.domain.model.mainfeed.MainFeedItem
+import com.jsmirabal.animeinfo.domain.model.mainfeed.MainFeedItemType
+
+data class DomainTopAnimes(
+    val topAnimes: List<AnimeCompact>,
+    override val feedItemType: MainFeedItemType
+) : MainFeedItem
