+++ app/src/main/java/com/jsmirabal/animeinfo/domain/model/mainfeed/DomainMainFeed.kt
@@ -1,0 +1,3 @@
+package com.jsmirabal.animeinfo.domain.model.mainfeed
+
+data class DomainMainFeed(val feedItems: List<MainFeedItem>)
