+++ app/src/main/java/com/jsmirabal/animeinfo/domain/model/mainfeed/MainFeedItem.kt
@@ -1,0 +1,5 @@
+package com.jsmirabal.animeinfo.domain.model.mainfeed
+
+interface MainFeedItem {
+    val feedItemType: MainFeedItemType
+}
