+++ app/src/main/java/com/jsmirabal/animeinfo/domain/model/mainfeed/MainFeedItemType.kt
@@ -1,0 +1,10 @@
+package com.jsmirabal.animeinfo.domain.model.mainfeed
+
+enum class MainFeedItemType {
+    RECOMMENDED,
+    CURRENT_SEASON,
+    TOP_AIRING_ANIMES,
+    TOP_UPCOMING_ANIMES,
+    MOST_POPULAR_ANIMES,
+    MOST_FAVORITE_ANIMES
+}
