+++ app/src/main/java/com/jsmirabal/animeinfo/domain/usecase/fetch/FetchAnimeDetailUseCase.kt
@@ -1,17 +1,17 @@
-package com.jsmirabal.animeinfo.domain.usecase
+package com.jsmirabal.animeinfo.domain.usecase.fetch
 
 import com.jsmirabal.animeinfo.domain.core.ResultWrapper
+import com.jsmirabal.animeinfo.domain.model.DomainAnimeDetail
 import com.jsmirabal.animeinfo.domain.model.DomainLayerError
-import com.jsmirabal.animeinfo.domain.model.DomainSeasonAnimes
 import com.jsmirabal.animeinfo.domain.repository.AnimeRepository
 import com.jsmirabal.animeinfo.domain.usecase.base.UseCase
 import kotlinx.coroutines.CoroutineScope
 
-class FetchCurrentSeasonUseCase(
+class FetchAnimeDetailUseCase(
     private val animeRepository: AnimeRepository,
     scope: CoroutineScope
-) : UseCase<Unit, ResultWrapper<DomainSeasonAnimes, DomainLayerError>>(scope) {
+) : UseCase<String, ResultWrapper<DomainAnimeDetail, DomainLayerError>>(scope) {
 
     @Suppress("PARAMETER_NAME_CHANGED_ON_OVERRIDE")
-    override suspend fun run(param: Unit) = animeRepository.fetchCurrentSeason()
+    override suspend fun run(animeId: String) = animeRepository.fetchAnimeDetail(animeId)
 }
