+++ app/src/main/java/com/jsmirabal/animeinfo/domain/usecase/fetch/FetchMostFavoriteAnimesUseCase.kt
@@ -1,4 +1,4 @@
-package com.jsmirabal.animeinfo.domain.usecase
+package com.jsmirabal.animeinfo.domain.usecase.fetch
 
 import com.jsmirabal.animeinfo.domain.core.ResultWrapper
 import com.jsmirabal.animeinfo.domain.model.DomainLayerError
@@ -7,11 +7,11 @@ import com.jsmirabal.animeinfo.domain.repository.AnimeRepository
 import com.jsmirabal.animeinfo.domain.usecase.base.UseCase
 import kotlinx.coroutines.CoroutineScope
 
-class FetchMostPopularAnimesUseCase(
+class FetchMostFavoriteAnimesUseCase(
     private val repository: AnimeRepository,
     scope: CoroutineScope
 ) : UseCase<String, ResultWrapper<DomainTopAnimes, DomainLayerError>>(scope) {
 
     @Suppress("PARAMETER_NAME_CHANGED_ON_OVERRIDE")
-    override suspend fun run(page: String) = repository.fetchMostPopularAnimes(page)
+    override suspend fun run(page: String) = repository.fetchMostFavoriteAnimes(page)
 }
