+++ app/src/test/java/com/jsmirabal/animeinfo/domain/usecase/base/UseCaseTest.kt
@@ -1,7 +1,6 @@
-package com.jsmirabal.animeinfo.domain.usecase
+package com.jsmirabal.animeinfo.domain.usecase.base
 
 import com.jsmirabal.animeinfo.TestDispatcherExtension
-import com.jsmirabal.animeinfo.domain.usecase.base.UseCase
 import io.mockk.clearAllMocks
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
