+++ app/src/test/java/com/jsmirabal/animeinfo/domain/usecase/fetch/FetchTopUpcomingAnimesUseCaseTest.kt
@@ -1,8 +1,8 @@
-package com.jsmirabal.animeinfo.domain.usecase
+package com.jsmirabal.animeinfo.domain.usecase.fetch
 
-import com.jsmirabal.animeinfo.data.ANIME_ID
+import com.jsmirabal.animeinfo.data.PAGE_NUMBER
 import com.jsmirabal.animeinfo.data.TestLogger
-import com.jsmirabal.animeinfo.data.dummyDomainAnimeVideosResultSuccess
+import com.jsmirabal.animeinfo.data.dummyDomainTopAnimesResultSuccess
 import com.jsmirabal.animeinfo.domain.repository.AnimeRepository
 import io.mockk.clearAllMocks
 import io.mockk.coEvery
@@ -17,10 +17,10 @@ import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Test
 
 @ExperimentalCoroutinesApi
-internal class FetchAnimeVideosUseCaseTest {
+internal class FetchTopUpcomingAnimesUseCaseTest {
     private val repository = mockk<AnimeRepository>()
     private val testScope = TestCoroutineScope()
-    private val useCase = FetchAnimeVideosUseCase(repository, testScope)
+    private val useCase = FetchTopUpcomingAnimesUseCase(repository, testScope)
 
     @AfterEach
     fun afterEach() {
@@ -30,16 +30,16 @@ internal class FetchAnimeVideosUseCaseTest {
     @Test
     fun run() = runBlockingTest {
         TestLogger.given("AnimeRepository returns mocked data")
-        coEvery { repository.fetchAnimeVideos(ANIME_ID) } returns dummyDomainAnimeVideosResultSuccess
+        coEvery { repository.fetchTopUpcomingAnimes(PAGE_NUMBER) } returns dummyDomainTopAnimesResultSuccess
 
-        TestLogger.whenever("FetchAnimeVideoUseCase is executed")
-        val result = useCase.run(ANIME_ID)
+        TestLogger.whenever("FetchTopUpcomingAnimesUseCase is executed")
+        val result = useCase.run(PAGE_NUMBER)
 
         TestLogger.then("Validate that the expected data is returned")
-        result shouldEqual dummyDomainAnimeVideosResultSuccess
+        result shouldEqual dummyDomainTopAnimesResultSuccess
 
-        TestLogger.then("Validate AnimeRepository#fetchAnimeVideo() was called")
-        coVerify { repository.fetchAnimeVideos(ANIME_ID) }
+        TestLogger.then("Validate AnimeRepository#fetchTopItems() was called")
+        coVerify { repository.fetchTopUpcomingAnimes(PAGE_NUMBER) }
 
         TestLogger.finally("Validate every method called from AnimeRepository was verified")
         confirmVerified(repository)
